refactor: 패키지명 application을 service로 이름 변경
