+++ src/main/java/wooteco/subway/ui/PathController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.application.PathService;
 import wooteco.subway.dto.PathResponse;
+import wooteco.subway.service.PathService;
 
 @RestController
 @RequestMapping("/paths")
