+++ src/test/java/wooteco/subway/ui/StationControllerTest.java
@@ -18,9 +18,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.application.StationService;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
+import wooteco.subway.service.StationService;
 
 @SpringBootTest
 @AutoConfigureMockMvc
